Move available_logical_cores to chia.util.cpu
